Visually documenting the TCRM workflow.
Intended that this will soon feed back into the main documentation,
and could also help facilitate future refactors of the code.

The visualisation and introspection should be an asset for
training new end users (and developers). The interactive
notebook facilitates interrupting any stage of the TCRM data flow
and inspecting or manipulating the intermediate products.